new update

+++ admin/benna_categories/report.php
@@ -29,20 +29,19 @@
                 <div class="card">
 
                     <div class="card-header">
-                        <button type="button" class="btn btn-primary waves-effect btn-sm" data-toggle="modal"
-                            data-target="#add-Modal"> أضافة قسم جديد <i class="fa fa-plus"></i> </button>
+                        <button type="button" class="btn btn-primary waves-effect btn-sm" data-toggle="modal" data-target="#add-Modal"> أضافة قسم جديد <i class="fa fa-plus"></i> </button>
                     </div>
                     <?php
                     if (isset($_SESSION['success_message'])) {
                         $message = $_SESSION['success_message'];
                         unset($_SESSION['success_message']);
-                        ?>
+                    ?>
                         <?php
                         ?>
                         <script src="plugins/jquery/jquery.min.js"></script>
                         <script src="plugins/sweetalert2/sweetalert2.min.js"></script>
                         <script>
-                            $(function () {
+                            $(function() {
                                 Swal.fire({
                                     position: 'center',
                                     icon: 'success',
@@ -56,12 +55,12 @@
                     } elseif (isset($_SESSION['error_messages'])) {
                         $formerror = $_SESSION['error_messages'];
                         foreach ($formerror as $error) {
-                            ?>
+                        ?>
                             <div class="alert alert-danger alert-dismissible" style="max-width: 800px; margin:20px">
                                 <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
                                 <?php echo $error; ?>
                             </div>
-                            <?php
+                    <?php
                         }
                         unset($_SESSION['error_messages']);
                     }
@@ -86,7 +85,7 @@
                                     $i = 0;
                                     foreach ($allcat as $cat) {
                                         $i++;
-                                        ?>
+                                    ?>
                                         <tr>
                                             <td>
                                                 <?php echo $i; ?>
@@ -94,71 +93,63 @@
                                             <td>
                                                 <?php echo $cat['name']; ?>
                                             </td>
-                                            <td> <img style="width: 60px; height:60px"
-                                                    src="benna_categories/images/<?php echo $cat['main_image']; ?> " alt="">
+                                            <td> <img style="width: 60px; height:60px" src="benna_categories/images/<?php echo $cat['main_image']; ?> " alt="">
                                             </td>
                                             <td>
-                                                <button type="button" class="btn btn-success btn-sm waves-effect"
-                                                    data-toggle="modal" data-target="#edit-Modal_<?php echo $cat['id']; ?>">
+                                                <button type="button" class="btn btn-success btn-sm waves-effect" data-toggle="modal" data-target="#edit-Modal_<?php echo $cat['id']; ?>">
                                                     تعديل <i class='fa fa-pen'></i> </button>
-                                                <a href="main.php?dir=categories&page=delete&cat_id=<?php echo $cat['id']; ?>"
-                                                    class="confirm btn btn-danger btn-sm"> حذف <i class='fa fa-trash'></i>
+                                                <a href="main.php?dir=categories&page=delete&cat_id=<?php echo $cat['id']; ?>" class="confirm btn btn-danger btn-sm"> حذف <i class='fa fa-trash'></i>
                                                 </a>
                                             </td>
                                         </tr>
                                         <!-- EDIT NEW CATEGORY MODAL   -->
-                                        <div class="modal fade" id="edit-Modal_<?php echo $cat['id']; ?>" tabindex="-1"
-                                            role="dialog">
+                                        <div class="modal fade" id="edit-Modal_<?php echo $cat['id']; ?>" tabindex="-1" role="dialog">
                                             <div class="modal-dialog" role="document">
                                                 <div class="modal-content">
                                                     <div class="modal-header">
                                                         <h4 class="modal-title"> تعديل القسم </h4>
                                                     </div>
-                                                    <form method="post" action="main.php?dir=benna_categories&page=edit"
-                                                        enctype="multipart/form-data">
+                                                    <form method="post" action="main.php?dir=benna_categories&page=edit" enctype="multipart/form-data">
                                                         <div class="modal-body">
                                                             <div class="form-group">
-                                                                <input type='hidden' name="cat_id"
-                                                                    value="<?php echo $cat['id']; ?>">
+                                                                <input type='hidden' name="cat_id" value="<?php echo $cat['id']; ?>">
                                                                 <label for="Company-2" class="block">الأسم </label>
-                                                                <input id="Company-2" required name="name" type="text"
-                                                                    class="form-control required"
-                                                                    value="<?php echo $cat['name'] ?>">
+                                                                <input id="Company-2" required name="name" type="text" class="form-control required" value="<?php echo $cat['name'] ?>">
                                                             </div>
                                                             <div class="form-group">
                                                                 <label for="Company-2" class="block"> الوصف </label>
-                                                                <textarea style="height: 150px;" id="Company-2"
-                                                                    name="description"
-                                                                    class="form-control"><?php echo $cat['description'] ?></textarea>
+                                                                <textarea style="height: 150px;" id="Company-2" name="description" class="form-control"><?php echo $cat['description'] ?></textarea>
                                                             </div>
                                                             <div class="form-group">
                                                                 <label for="Company-2" class="block"> الوصف المختصر </label>
-                                                                <textarea style="height: 150px;" id="Company-2"
-                                                                    name="short_description"
-                                                                    class="form-control"><?php echo $cat['short_description'] ?></textarea>
+                                                                <textarea style="height: 150px;" id="Company-2" name="short_description" class="form-control"><?php echo $cat['short_description'] ?></textarea>
                                                             </div>
                                                             <div class="form-group">
                                                                 <label for="customFile"> تعديل صورة القسم </label>
                                                                 <div class="custom-file">
-                                                                    <input type="file" class="custom-file-input"
-                                                                        id="customFile" accept='image/*' name="main_image">
+                                                                    <input type="file" class="custom-file-input" id="customFile" accept='image/*' name="main_image">
                                                                     <label class="custom-file-label" for="customFile">اختر
                                                                         الصورة</label>
                                                                 </div>
                                                             </div>
+                                                            <div class="form-group">
+                                                                <label for="customFile"> تعديل صورة بانر القسم </label>
+                                                                <div class="custom-file">
+                                                                    <input required type="file" class="custom-file-input" id="customFile" accept='image/*' name="main_image_banner">
+                                                                    <label class="custom-file-label" for="customFile">اختر الصورة</label>
+                                                                </div>
+                                                            </div>
                                                         </div>
                                                         <div class="modal-footer">
-                                                            <button type="submit" name="edit_cat"
-                                                                class="btn btn-primary waves-effect waves-light "> تعديل
+                                                            <button type="submit" name="edit_cat" class="btn btn-primary waves-effect waves-light "> تعديل
                                                             </button>
-                                                            <button type="button" class="btn btn-default waves-effect "
-                                                                data-dismiss="modal">رجوع</button>
+                                                            <button type="button" class="btn btn-default waves-effect " data-dismiss="modal">رجوع</button>
                                                         </div>
                                                     </form>
                                                 </div>
                                             </div>
                                         </div>
-                                        <?php
+                                    <?php
                                     }
                                     ?>
                             </table>
@@ -173,37 +164,38 @@
                             <div class="modal-header">
                                 <h4 class="modal-title">أضافة قسم رئيسي </h4>
                             </div>
-                            <form action="main.php?dir=benna_categories&page=add" method="post"
-                                enctype="multipart/form-data">
+                            <form action="main.php?dir=benna_categories&page=add" method="post" enctype="multipart/form-data">
                                 <div class="modal-body">
                                     <div class="form-group">
                                         <label for="Company-2" class="block"> الأسم </label>
-                                        <input required id="Company-2" name="name" type="text"
-                                            class="form-control required">
+                                        <input required id="Company-2" name="name" type="text" class="form-control required">
                                     </div>
                                     <div class="form-group">
                                         <label for="Company-2" class="block"> الوصف </label>
-                                        <textarea style="height: 150px;" id="Company-2" name="description"
-                                            class="form-control"></textarea>
+                                        <textarea style="height: 150px;" id="Company-2" name="description" class="form-control"></textarea>
                                     </div>
 
                                     <div class="form-group">
                                         <label for="Company-2" class="block"> الوصف المختصر </label>
-                                        <textarea style="height: 100px;" id="Company-2" name="short_description"
-                                            class="form-control"></textarea>
+                                        <textarea style="height: 100px;" id="Company-2" name="short_description" class="form-control"></textarea>
                                     </div>
                                     <div class="form-group">
                                         <label for="customFile"> صورة القسم </label>
                                         <div class="custom-file">
-                                            <input required type="file" class="custom-file-input" id="customFile"
-                                                accept='image/*' name="main_image">
+                                            <input required type="file" class="custom-file-input" id="customFile" accept='image/*' name="main_image">
+                                            <label class="custom-file-label" for="customFile">اختر الصورة</label>
+                                        </div>
+                                    </div>
+                                    <div class="form-group">
+                                        <label for="customFile"> صورة بانر القسم </label>
+                                        <div class="custom-file">
+                                            <input required type="file" class="custom-file-input" id="customFile" accept='image/*' name="main_image_banner">
                                             <label class="custom-file-label" for="customFile">اختر الصورة</label>
                                         </div>
                                     </div>
                                 </div>
                                 <div class="modal-footer">
-                                    <button type="submit" name="add_cat"
-                                        class="btn btn-primary waves-effect waves-light "> حفظ </button>
+                                    <button type="submit" name="add_cat" class="btn btn-primary waves-effect waves-light "> حفظ </button>
                                     <button type="button" class="btn btn-default waves-effect " data-dismiss="modal">
                                         رجوع </button>
